requirements.txt
llama-cpp-python
yfinance
PyYAML
LangChain
langchain-community
langchain-core
langchain_experimental
sentence-transformers
requests